unify mutate and scan operations to simplify higher order plumbing

+++ csrc/exec.h
@@ -40,7 +40,7 @@ static inline void extract(vector dest, vector keys, value *r)
 
 static inline void store(value *r, value reg, value v)
 {
-    if ((unsigned long)reg != register_ignore)
+    if (reg != register_ignore)
         r[toreg(reg)] = v;
 }
